refactor: game console 구조 변경
