+++ apps/web/src/app/[locale]/landing/RankingSection/GameConsole/SvgContainer.tsx
